+++ CHANGELOG.md
@@ -1,3 +1,80 @@
+# [1.2.0-next.10](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.9...v1.2.0-next.10) (2023-06-04)
+
+
+### Features
+
+* add allow generators for func-names ([9674cb3](https://github.com/vi-latyshev/eslint-config/commit/9674cb3b93b6bb8c444d8b1a44cf2937c3f5d5b5))
+
+# [1.2.0-next.9](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.8...v1.2.0-next.9) (2023-06-04)
+
+
+### Features
+
+* add consistent imports ([7970aad](https://github.com/vi-latyshev/eslint-config/commit/7970aad1a12e594a1f84e18b9b9412663d2052e8))
+
+# [1.2.0-next.8](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.7...v1.2.0-next.8) (2023-05-17)
+
+
+### Bug Fixes
+
+* import order for alias ([215af34](https://github.com/vi-latyshev/eslint-config/commit/215af344c3231c79f83ba225742a71ea4c283377))
+
+# [1.2.0-next.7](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.6...v1.2.0-next.7) (2023-05-08)
+
+
+### Bug Fixes
+
+* no-unused-vars for js files ([56a236e](https://github.com/vi-latyshev/eslint-config/commit/56a236e6b364dc4e6743001c4d4fed1cbdd776e4))
+
+# [1.2.0-next.6](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.5...v1.2.0-next.6) (2023-05-08)
+
+
+### Bug Fixes
+
+* always lines between class ([92fe944](https://github.com/vi-latyshev/eslint-config/commit/92fe944b150651875267fd80a449de919d55cdfc))
+
+# [1.2.0-next.5](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.4...v1.2.0-next.5) (2023-05-07)
+
+
+### Bug Fixes
+
+* next config & react in jsx scope & pathGroup in import/orde ([e95d036](https://github.com/vi-latyshev/eslint-config/commit/e95d036f25208fa081706d6bc97cb992cbece724))
+
+# [1.2.0-next.4](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.3...v1.2.0-next.4) (2023-05-07)
+
+
+### Bug Fixes
+
+* **react:** jsx-filename-extension for tsx ([b82f369](https://github.com/vi-latyshev/eslint-config/commit/b82f36913a5f157e2ca1f1a9e51dacff6adbd1cc))
+
+# [1.2.0-next.3](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.2...v1.2.0-next.3) (2023-05-07)
+
+
+### Bug Fixes
+
+* **react:** fix pages for prefer default for next.js ([116f824](https://github.com/vi-latyshev/eslint-config/commit/116f824bf97f3621e98d39befed4518689b55689))
+
+# [1.2.0-next.2](https://github.com/vi-latyshev/eslint-config/compare/v1.2.0-next.1...v1.2.0-next.2) (2023-05-07)
+
+
+### Features
+
+* **react:** add pages for prefer default for next.js ([9fac948](https://github.com/vi-latyshev/eslint-config/commit/9fac948a089efe7417e8c5cfe7718c746a66ae43))
+
+# [1.2.0-next.1](https://github.com/vi-latyshev/eslint-config/compare/v1.1.1-next.1...v1.2.0-next.1) (2023-05-06)
+
+
+### Features
+
+* prefer default import for next.js app router ([220cc08](https://github.com/vi-latyshev/eslint-config/commit/220cc08001322f152a7f180ebf7eec4eecf19164))
+
+## [1.1.1-next.1](https://github.com/vi-latyshev/eslint-config/compare/v1.1.0...v1.1.1-next.1) (2023-05-06)
+
+
+### Bug Fixes
+
+* dependencies & fully next for react ([e5c14d2](https://github.com/vi-latyshev/eslint-config/commit/e5c14d2f5acef26042965f33e34d1ca8c8d85e64))
+
 # [1.1.0](https://github.com/vi-latyshev/eslint-config/compare/v1.0.3...v1.1.0) (2023-05-06)